+++ agent-lib/src/test/java/local/wspolnyprojekt/nodeagentlib/AgentRestRequestDetailsGitTest.java
@@ -1,12 +1,12 @@
-package local.wspolnyprojekt.nodeagentlib.common.requestcreator;
+package local.wspolnyprojekt.nodeagentlib;
 
-import local.wspolnyprojekt.nodeagentlib.common.GitCredentials;
-import local.wspolnyprojekt.nodeagentlib.common.RequestDetails;
+import local.wspolnyprojekt.nodeagentlib.dto.GitCredentials;
+import local.wspolnyprojekt.nodeagentlib.dto.RequestDetails;
+import local.wspolnyprojekt.nodeagentlib.dto.RestEndpoints;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.web.bind.annotation.RequestMethod;
 
-import static local.wspolnyprojekt.nodeagentlib.common.RestEndpoints.*;
 import static local.wspolnyprojekt.nodeagentlib.AgentRestRequestDetails.*;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
@@ -24,7 +24,7 @@ class AgentRestRequestDetailsGitTest {
         GitCredentials credentials = new GitCredentials(username, password);
         RequestDetails requestDetails = gitCredentialsRequestDetails(username, password);
         assertThat(requestDetails.getRequestMethod()).isEqualTo(RequestMethod.POST);
-        assertThat(requestDetails.getUriEndpoint()).isEqualTo(GIT_CREDENTIALS_ENDPOINT);
+        assertThat(requestDetails.getUriEndpoint()).isEqualTo(RestEndpoints.GIT_CREDENTIALS_ENDPOINT);
         assertThat(requestDetails.getJsonPayload()).isEqualTo(credentials.getJsonString());
     }
 
@@ -34,7 +34,7 @@ class AgentRestRequestDetailsGitTest {
         String branch = "branch";
         RequestDetails requestDetails = gitCloneRequestDetails(repository, branch, taskId);
         assertThat(requestDetails.getRequestMethod()).isEqualTo(RequestMethod.POST);
-        assertThat(requestDetails.getUriEndpoint()).isEqualTo(GIT_ENDPOINT.replace("{" + TASKID_PATH_VARIABLE + "}", taskId));
+        assertThat(requestDetails.getUriEndpoint()).isEqualTo(RestEndpoints.GIT_ENDPOINT.replace("{" + RestEndpoints.TASKID_PATH_VARIABLE + "}", taskId));
         assertThat(requestDetails.getJsonPayload()).isEqualTo(String.format("{\"repositoryUrl\":\"%s\",\"branch\":\"%s\"}", repository, branch));
     }
 
@@ -43,7 +43,7 @@ class AgentRestRequestDetailsGitTest {
         RequestDetails requestDetails = gitPullRequestDetails(taskId);
         assertThat(requestDetails.getRequestMethod()).isEqualTo(RequestMethod.PUT);
         assertThat(requestDetails.getJsonPayload()).isEmpty();
-        assertThat(requestDetails.getUriEndpoint()).isEqualTo(GIT_ENDPOINT.replace("{" + TASKID_PATH_VARIABLE + "}", taskId));
+        assertThat(requestDetails.getUriEndpoint()).isEqualTo(RestEndpoints.GIT_ENDPOINT.replace("{" + RestEndpoints.TASKID_PATH_VARIABLE + "}", taskId));
     }
 
     @Test
@@ -51,6 +51,6 @@ class AgentRestRequestDetailsGitTest {
         RequestDetails requestDetails = getSystemLoadRequestDetails();
         assertThat(requestDetails.getRequestMethod()).isEqualTo(RequestMethod.GET);
         assertThat(requestDetails.getJsonPayload()).isEmpty();
-        assertThat(requestDetails.getUriEndpoint()).isEqualTo(SYSTEM_LOAD);
+        assertThat(requestDetails.getUriEndpoint()).isEqualTo(RestEndpoints.SYSTEM_LOAD);
     }
 }
