+++ agent/src/main/java/local/wspolnyprojekt/nodeagent/git/Credentials.java
@@ -1,6 +1,6 @@
 package local.wspolnyprojekt.nodeagent.git;
 
-import local.wspolnyprojekt.nodeagentlib.common.GitCredentials;
+import local.wspolnyprojekt.nodeagentlib.dto.GitCredentials;
 import lombok.Getter;
 import org.eclipse.jgit.transport.UsernamePasswordCredentialsProvider;
 import org.springframework.stereotype.Component;
