+++ agent/src/main/java/local/wspolnyprojekt/nodeagent/git/JGitClient.java
@@ -1,6 +1,6 @@
 package local.wspolnyprojekt.nodeagent.git;
 
-import local.wspolnyprojekt.nodeagentlib.common.GitResource;
+import local.wspolnyprojekt.nodeagentlib.dto.GitResource;
 import lombok.RequiredArgsConstructor;
 import org.eclipse.jgit.api.Git;
 import org.eclipse.jgit.api.errors.GitAPIException;
